Add files via upload
Aiming the working process of Arduino uno and PIR sensor, the project first deals with the sensitivity detector of people inside a room using PIR sensor, then the paper optimizes the state of Arduino Uno board . its future developments and applications in different fields. power consumption can be optimized and store it for further usage.Our society will get various benefits in the fulfilment of this project.